+++ lib/src/eval/compiler/statement/assert.dart
@@ -1,12 +1,10 @@
 import 'package:analyzer/dart/ast/ast.dart';
-import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:dart_eval/src/eval/compiler/builtins.dart';
 import 'package:dart_eval/src/eval/compiler/context.dart';
 import 'package:dart_eval/src/eval/compiler/expression/expression.dart';
+import 'package:dart_eval/src/eval/compiler/helpers/assert.dart';
 import 'package:dart_eval/src/eval/compiler/statement/statement.dart';
 import 'package:dart_eval/src/eval/compiler/type.dart';
-import 'package:dart_eval/src/eval/compiler/variable.dart';
-import 'package:dart_eval/src/eval/runtime/runtime.dart';
 
 StatementInfo compileAssertStatement(AssertStatement s, CompilerContext ctx,
     AlwaysReturnType? expectedReturnType) {
@@ -15,17 +13,7 @@ StatementInfo compileAssertStatement(AssertStatement s, CompilerContext ctx,
       ? compileExpression(s.message!, ctx)
       : BuiltinValue().push(ctx);
 
-  msg.pushArg(ctx);
-  ctx.pushOp(
-      InvokeExternal.make(ctx.bridgeStaticFunctionIndices[
-          ctx.libraryMap['dart:core']]!['AssertionError.']!),
-      InvokeExternal.LEN);
-  ctx.pushOp(PushReturnValue.make(), PushReturnValue.LEN);
-  final assertionErr = Variable.alloc(ctx,
-      TypeRef.fromBridgeTypeRef(ctx, BridgeTypeRef(CoreTypes.assertionError)));
-
-  ctx.pushOp(Assert.make(cond.scopeFrameOffset, assertionErr.scopeFrameOffset),
-      Assert.LEN);
+  doAssert(ctx, cond, msg);
 
   return StatementInfo(-1);
 }