+++ lib/stdlib/typed_data.dart
@@ -1,0 +1,2 @@
+/// Provides dart:typed_data bridge classes and wrappers
+library dart_eval.stdlib.typed_data;